Safari-609.1.11.1/LayoutTests/fast/hidpi/image-srcset-data-escaped-srcset-expected.txt

PASS document.getElementById("foo").clientWidth==50 is true
This test passes if the image below is not empty. It ensures that the srcset attribute support data URI schemes with escaped characters.
